+++ families/mc1201/common/src/main/java/dev/foucaultleon/flterraforged/api/mc1201/materializer/DelegatingBlockMaterializer.java
@@ -169,4 +169,9 @@ public abstract class DelegatingBlockMaterializer implements BlockMaterializer {
     public int hydrologyBankSealDepth() {
         return delegate.hydrologyBankSealDepth();
     }
+
+    @Override
+    public void decorateWatercourses(WaterDecorationContext context) {
+        delegate.decorateWatercourses(context);
+    }
 }